+++ app/src/main/java/opensource/cached_dupe_scanner/tasks/TaskCoordinator.kt
@@ -20,6 +20,7 @@ enum class TaskKind {
     RebuildGroups,
     ClearCache,
     EmptyTrash,
+    BulkDelete,
     SimilarityExperiment
 }